+++ packages/plugin-core/src/test/utils/vscode-test-utils.ts
